rewrote the code for tests, added Gson, added test for Gson and Json parser

+++ app/src/main/java/piftik/github/com/weatherproject/utils/TemperatureConverter.java
@@ -1,7 +1,7 @@
 package piftik.github.com.weatherproject.utils;
 
 public class TemperatureConverter {
-    public static int convertTemperatureToCelsius(int kelvinTemperature) {
+    public static double convertTemperatureToCelsius(double kelvinTemperature) {
         return kelvinTemperature - 273;
     }
 }
